backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.116.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Telegram
python-telegram-bot==20.7

# Email Parsing
email-validator==2.1.0.post1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# CORS
starlette==0.35.1
